+++ lib/CreationCompte/PremierePage/deuxiemeUtilisation.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 
-import '../../utilisateur.dart';
+import '../../heureDisponible.dart';
+import '../../main.dart';
 
 class DeuxiemeUtilisation extends StatefulWidget {
   @override
@@ -42,13 +43,75 @@ Text(
  disponibilite(jour.elementAt(4)),
  disponibilite(jour.elementAt(5)),
  disponibilite(jour.elementAt(6)),
+ Row(
+                        mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                        children: <Widget>[
+                          RaisedButton(
+                              child: Text('Retour'),
+                              onPressed: () {
+                                Navigator.pop(context);}),
+                          RaisedButton(
+                              child: Text('Reset'),
+                              onPressed: () {
+                               setState(() {
+                                 utilisateur.disponibiliteSemaine.clear();   
+                               });
+                                
  
+                                
+                               
+                              }),      
+                          RaisedButton(
+                              child: Text('Confirmer'),
+                              onPressed: () {
+                               
+                                   
+ 
+                                Navigator.pushReplacementNamed(
+                                    context, '/home');
+                               
+                              }),
+                        ]),
+ Container(
+                     padding: EdgeInsets.fromLTRB(50, 100, 50, 0),
+                     height: 400,
+                     
+                     child:Row(  mainAxisAlignment: MainAxisAlignment.spaceBetween, children: <Widget>[
+                       getTextWidgets(utilisateur.disponibiliteSemaine),
+getTextWidgets2(utilisateur.disponibiliteSemaine),
+getTextWidgets3(utilisateur.disponibiliteSemaine),
+                     ], )
+                       
+                    ) 
     ],
   )
 ),
         )));
   }
-
+ Widget getTextWidgets(List<HeureDisponible> strings)
+  {
+    List<Widget> list = new List<Widget>();
+    for(var i = 0; i < strings.length; i++){
+        list.add(new Text(strings.elementAt(i).jour));
+           }
+    return new Column( children: list);
+  } 
+   Widget getTextWidgets2(List<HeureDisponible> strings)
+  {
+    List<Widget> list = new List<Widget>();
+    for(var i = 0; i < strings.length; i++){
+        list.add(new Text('Début :'+strings.elementAt(i).debut.toString().substring(10,15)));
+           }
+    return new Column( children: list);
+  } 
+   Widget getTextWidgets3(List<HeureDisponible> strings)
+  {
+    List<Widget> list = new List<Widget>();
+    for(var i = 0; i < strings.length; i++){
+        list.add(new Text('Fin :'+strings.elementAt(i).fin.toString().substring(10,15)));
+           }
+    return new Column( children: list);
+  } 
    Widget disponibilite (String jour){
 return Row(
   mainAxisAlignment: MainAxisAlignment.spaceBetween,
